[Nokia][devicedata] Modified the port autoneg default setting for Nokia-7215 platform (#13459)
Why I did it
autoneg is not supported in the previous release 202012 on Nokia-7215 platform. To migrate to the 202205 image with autoneg support, we need to disable the autoneg to allow the link to be up when issue load minigraph. This requires to change the autoneg setting to be off in the port_config.ini file.

How I did it
Modify the port_config.ini to set the autoneg off.

How to verify it
Running the new image, with load mingraph.xml, execute "show int autonet status" should show autoneg disabled