[marvell-armhf][uboot] Fixed the uboot setting for sonic-installer set-default form 202012 to 202205 branch. (#13911)
#### Why I did it
When using ```sonic-install set-default``` to switch the image from 202012 to 202205.  The system will be stuck at loading kernel while reboot.

#### How I did it
The issue is caused by the kernal size related setting in uboot environment is smaller in the 202012 branch while they are larger in 202205 branch.  The "sonic-installer set-default" just changes the boot_next variable.  To fix this issue, we sync up the 202012 branch kernel related setting with the 202205 branch.  This PR is only applicable to 202012 branch.

#### How to verify it
1) Install the latest 202205 image .89 or latest and reboot
2) Install the 202012 image which contains this fix and reboot
3) using "sonic-installer set-default 202205 image and reboot
4) system should start without any issue. 
  
#### Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.