Refactor the logic of tagging kube container as local latest (#14367)
Why I did it
We found a bug when pilot, the tag function doesn't remove the ACR domain when do tag, it makes the latest tag not work. And in the original tag function, it calls os.system and os.popen which are not recommend, need to refactor.

How I did it
Do a split("/") when get image_rep to fix the acr domain bug
Refactor the tag function code and add test cases

How to verify it
Check whether container images are tagged as latest when in kube mode.