[sonic-frr] enable info log for graceful restart events (#15623)
#### Why I did it
cherry pick, #15535

Graceful restart is a key event for bgpd, related log print is debug level. To change it to info level to get more visibilities when this kind of event is triggered.

##### Work item tracking
- Microsoft ADO **(13875291)**:

#### How I did it
To create patch file to change from debug level to info level. 

#### How to verify it
To run PR test and capture the print.