JIRA-ONSBUSWI-2698: [ec-sonic][common driver][pmbus][ym2651y.c] When psu_power_good is failed, temperature return 1023 degree.
[why to do]:
The driver will poll for 3 seconds at a time to update the pmbus value.
Before modification, psu_power_good = failed ==>The temperature will be returned as 0 (degrees).
This will cause the sonic to send syslog when the psu is powered up, but detect 0 degrees before updating the correct temperature. (misjudgment)
Therefore, it is necessary to modify the value that when psu_power_good = failed ==> the temperature returns 1023 (degrees), so that the python program can know that the status of 1023 is for power_good is failed value(befor update the correct temperature value).