[Platform][AS9817-32D/O] Support 10G/1G on management ports
There is a retimer between SFP ports and MAC chip, and preemphasis has
to be configured on this retimer.

This funcion is based on set_sfp_mux, which is a platform api of sfp

Note:
xcvrd listents to config DB and call the platform api to set the correct
configuration on the re-timer.
Refer to the following PR.
    * https://github.com/edge-core/sonic-buildimage/pull/320
    * https://github.com/edge-core/sonic-platform-daemons/pull/17