SON-2218: es9618xx: fix platform deb package internal build error detection
Fix internal error detection during the build process of the platform deb package.
Before this fix, the build process completed successfully even if there was an error
creating the platform deb package. After this fix, the process correctly fails when
an error occurs with the specific platform deb package name.

Signed-off-by: Asher Oved <ashero@xsightlabs.com>