SON-2263: es9618xx: Fix VDDH value on boot for R0B
For R0B boxes, the VDDH voltage must be increased by 50 mV
for SerDes running the new firmware to operate correctly.

This change fixes a regression that was introduced when
porting the functionality from XsightOS to SONiC in the
"SON-2263: es9618xx: Configure VDDH in R0B on boot" commit.

Signed-off-by: Michael Zaidman <michaelz@xsightlabs.com>